Maricel Cart 04-19-22 07:21:00

+++ Pick-n-Pay-Store/Models/Model Props /Billings+CoreDataProperties.swift
@@ -20,7 +20,7 @@ extension Billings {
     @NSManaged public var firstName: String?
     @NSManaged public var lastName: String?
     @NSManaged public var billingAddress: String?
-    @NSManaged public var isDefault: Int
+    @NSManaged public var isDefault: Int16
 
 }
